+++ src/roboto/version/autogen_version.py
@@ -1,4 +1,4 @@
-AUTOGEN_VERSION = "0.28.2"
+AUTOGEN_VERSION = "0.29.0rc0"
 
 __all__= ("AUTOGEN_VERSION",)